Added Unity Project to Repo
This was a game that was made for my Craft of Game Development coursework.
I moved it to a public repo because the original repo is private and I don't own it, my university does, so I was unable to set it to public.
I decided that instead of giving up, I would instead bring it here as a technical work-around so that you can still see and have access to the 2D Platformer game I made.
Thank you for your time, and apologies for the inconvenience!